[server] Extract RequextContext from EnhancedLogContext
